extract ConversationEvent
